% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rSharkClient.R
\name{rShark.createLogicalLists}
\alias{rShark.createLogicalLists}
\title{Creates list of logical lists Function}
\usage{
rShark.createLogicalLists(logLists)
}
\arguments{
\item{logLists}{list of vectors of strings.}
}
\value{
List<List>
}
\description{
This function creates List of type List on backend.
}
\examples{
rShark.createLogicalLists (list(c("ProcessMetrics","ProductMetrics"), c("ID","Name"))
}
\keyword{createLogicalLists}

