% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rSharkClient.R
\name{rShark.createMongoDBUtils}
\alias{rShark.createMongoDBUtils}
\title{Create MongoDBUtils Function}
\usage{
rShark.createMongoDBUtils(sparkSession)
}
\arguments{
\item{SparkSession}{Provided from outside with specific parameters.}
}
\value{
MongoDBUtils
}
\description{
This function creates MongoDBUtils on backend.
}
\examples{
rShark.createMongoDBUtils(SparkSession)
}
\keyword{mongodbutils}

