% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rSharkClient.R
\name{rShark.loadDataLogical}
\alias{rShark.loadDataLogical}
\title{Call loadDataLogical Function}
\usage{
rShark.loadDataLogical(mongoDBUtils, collectionName, logicalLists)
}
\arguments{
\item{mongoDBUtils}{To be provided on which loadDataLogical method is called.}

\item{collectionName}{To be provided for which data is retrived from backend.}

\item{logicalLists}{List of type List to provided on which filtering is performed.}
}
\value{
SparkDataFrame
}
\description{
This function calls loadDataLogical on backend.
}
\examples{
rShark.loadDataLogical (mongoDBUtils,collectionName,logicalLists)
}
\keyword{loadDataLogical}

