% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rSharkClient.R
\name{rShark.writeData}
\alias{rShark.writeData}
\title{Call writeData Function}
\usage{
rShark.writeData(mongoDBUtils, dataFrame, collectionName)
}
\arguments{
\item{mongoDBUtils}{To be provided on which writeData method is called.}

\item{dataFrame}{To be provided which is upserted in the Database.}

\item{collectionName}{To be provided where Data will be upserted.}
}
\description{
This function calls writeData on backend.
}
\examples{
rShark.writeData (mongoDBUtils,dataFrame,collectionName)
}
\keyword{writeData}

