% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rSharkClient.R
\name{rShark.loadData}
\alias{rShark.loadData}
\title{Call loadData Function}
\usage{
rShark.loadData(mongoDBUtils, collectionName)
}
\arguments{
\item{mongoDBUtils}{To be provided on which loadData method is called.}

\item{collectionName}{To be provided for which data is retrived from backend.}
}
\value{
SparkDataFrame
}
\description{
This function calls loadData on backend.
}
\examples{
rShark.loadData (mongoDBUtils,collectionName)
}
\keyword{loadData}

