% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rSharkClient.R
\name{rShark.loadDataLogical}
\alias{rShark.loadDataLogical}
\title{Call loadDataLogical Function}
\usage{
rShark.loadDataLogical(mongoDBUtils, collectionName, logLists)
}
\arguments{
\item{mongoDBUtils}{To be provided on which loadDataLogical method is called.}

\item{collectionName}{To be provided for which data is retrived from backend.}

\item{logLists}{List of vectors of strings.}
}
\value{
SparkDataFrame
}
\description{
This function calls loadDataLogical on backend.
}
\examples{
rShark.loadDataLogical (mongoDBUtils,collectionName,list(c("ProcessMetrics","ProductMetrics"), c("ID","Name")))
}
\keyword{loadDataLogical}

